+++ pubspec.yaml
@@ -35,3 +35,9 @@ flutter:
           weight: 700
         - asset: assets/fonts/NanumSquareEB.ttf
           weight: 900
+    - family: MaruBuri
+      fonts:
+        - asset: assets/fonts/MaruBuri-Regular.ttf
+    - family: NanumGaRamYeonGgoc
+      fonts:
+        - asset: assets/fonts/NanumGaRamYeonGgoc.ttf
